对比协议修改

+++ src/main/java/com/example/moviesapi/Entity/BoxOffice_revenue.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class BoxOffice_revenue {
-    private Integer worldwide;
-    private Integer openingWeekendUSA;
+    private Long worldwide;
+    private Long openingWeekendUSA;
 }
